Move `boolean/equals` to `value/equals` This commit moves `boolean/equals` to `value/equals` to ensure predictability of the input types.
The commit also fixes a typo in the `equals()` specification.
