+++ src/value/equals.js
